+++ vm/actor/src/util/chaos/state.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use crate::UnmarshallableCBOR;
+use crate::util::unmarshallable::UnmarshallableCBOR;
 
 pub struct State {
     pub unmarshallable: Vec<UnmarshallableCBOR>,
