+++ vm/actor/src/util/mod.rs
@@ -15,4 +15,3 @@ pub use self::balance_table::BalanceTable;
 pub use self::multimap::*;
 pub use self::set::Set;
 pub use self::set_multimap::SetMultimap;
-pub use unmarshallable::*;
