+++ vm/actor/src/util/puppet/state.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use crate::UnmarshallableCBOR;
+use crate::util::unmarshallable::UnmarshallableCBOR;
 use encoding::{tuple::*, Cbor};
 
 #[derive(Default, Serialize_tuple, Deserialize_tuple)]
